refactor: move shared logger to core utilities (#97)
