+++ src/nip46/client.ts
@@ -5,7 +5,7 @@ import { encrypt as encryptNIP44, decrypt as decryptNIP44 } from "../nip44";
 import { NostrEvent, NostrFilter } from "../types/nostr";
 import { createSignedEvent } from "../nip01/event";
 import { parseConnectionString } from "./utils/connection";
-import { Logger, LogLevel } from "./utils/logger";
+import { Logger, LogLevel } from "../utils/logger";
 import { generateRequestId } from "./utils/request-response";
 import { isValidAuthUrl } from "./utils/auth";
 import {
